+++ tests/unit/layouts/HeaderLayout.spec.js
@@ -1,6 +1,6 @@
 import { shallowMount } from '@vue/test-utils'
 import HeaderLayout from '@/layouts/HeaderLayout.vue'
-import NavItem from '@/components/NavItem.vue'
+import NavItem from '@/layouts/components/NavItem.vue'
 
 describe('HeaderLayout', () => {
